Update files changed in branch optimization
While working on an branch optimization the branch matteo/Liouvillian
was merged in to optimization. Later the implementation of the
Lindbladian was changed in favor of a  Jordan-Wigner transformation.
Therefore the changed and added files for the optimization of the
auxiliary hybridization wherend updated to main.
They are needed now in this  branch.